measurements/requirements.txt
tensorflow
matplotlib
scikit-learn
numpy
fastapi
uvicorn
Pillow
python-multipart
